1. Started use of ON_EACH_CPU instead of on_each_cpu. ON_EACH_CPU is appropiately defined in seeker.h depending on the linux version.

+++ Module/seeker-scheduler/debug.c
@@ -97,7 +97,7 @@ void debug_free(struct debug_block *p)
 void purge_debug(void)
 {
 	struct debug_block *c1,*c2;
-	unsigned int flags;
+	unsigned long flags;
 	if(start_debug == NULL || current_debug == NULL)
 		return;
 	/* Acquire the lock, then set current debug to NULL
